rolling update st fix selectors intend and logging
